fix: #1 修复脚本超时问题

+++ upgrade-scripts/upgrade-22.3.7.php
@@ -98,7 +98,7 @@ class EduSohoUpgrade extends AbstractUpdater
     public function registerSyncTask($page)
     {
         if($page > 1) {
-            $logPage = $page / 1000;
+            $logPage = (int)($page / 1000);
             $coursePage = $page % 1000;
         }else {
             $logPage = 1;
